refactor :: excel service
