+++ src/main/java/com/keepgoing/keepserver/global/config/ExcelConfig.java
@@ -3,13 +3,13 @@ package com.keepgoing.keepserver.global.config;
 import com.keepgoing.keepserver.global.file.generate.GenerateExcelTemplate;
 import com.keepgoing.keepserver.global.file.generate.GenerateExcelTemplateImpl;
 import com.keepgoing.keepserver.global.file.mapper.RowMapper;
-import com.keepgoing.keepserver.global.file.process.ExcelProcessor;
+import com.keepgoing.keepserver.global.file.processor.ExcelProcessor;
 import com.keepgoing.keepserver.global.file.reader.ExcelReader;
 import com.keepgoing.keepserver.global.file.reader.ExcelReaderImpl;
-import com.keepgoing.keepserver.global.file.validate.ExcelValidator;
-import com.keepgoing.keepserver.domain.teacher.service.ItemDtoRowMapperImpl;
+import com.keepgoing.keepserver.global.file.validator.ExcelValidator;
+import com.keepgoing.keepserver.global.file.mapper.ItemDtoRowMapperImpl;
 import com.keepgoing.keepserver.domain.teacher.payload.ItemExcelDto;
-import com.keepgoing.keepserver.global.file.validate.ExcelValidatorImpl;
+import com.keepgoing.keepserver.global.file.validator.ExcelValidatorImpl;
 import jakarta.validation.Validator;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
