--- src/main/java/com/keepgoing/keepserver/global/file/validate/ExcelValidator.java
@@ -1,7 +1,0 @@
-package com.keepgoing.keepserver.global.file.validate;
-
-import java.util.List;
-
-public interface ExcelValidator<T> {
-    List<ExcelValidationResult<T>> validate(List<T> dtos);
-}
