+++ src/main/java/com/keepgoing/keepserver/global/file/validator/ExcelValidationErrorResponse.java
@@ -1,4 +1,4 @@
-package com.keepgoing.keepserver.domain.teacher.payload;
+package com.keepgoing.keepserver.global.file.validator;
 
 import java.util.List;
