+++ src/main/java/com/keepgoing/keepserver/global/file/validator/ExcelValidator.java
@@ -1,0 +1,9 @@
+package com.keepgoing.keepserver.global.file.validator;
+
+import com.keepgoing.keepserver.global.file.model.ExcelValidationResult;
+
+import java.util.List;
+
+public interface ExcelValidator<T> {
+    List<ExcelValidationResult<T>> validate(List<T> dtos);
+}
